Refactor: 오타 수정
