+++ src/main/java/com/ceos22/cgv_clone/domain/booking/repository/BookingRepository.java
@@ -1,11 +1,10 @@
 package com.ceos22.cgv_clone.domain.booking.repository;
 
-import com.ceos22.cgv_clone.domain.booking.eneity.Booking;
+import com.ceos22.cgv_clone.domain.booking.entity.Booking;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-import java.util.List;
 import java.util.Optional;
 
 public interface BookingRepository extends JpaRepository<Booking, Long> {
