+++ src/main/java/com/ceos22/cgv_clone/domain/booking/repository/BookingSeatRepository.java
@@ -1,7 +1,7 @@
 package com.ceos22.cgv_clone.domain.booking.repository;
 
-import com.ceos22.cgv_clone.domain.booking.eneity.Booking;
-import com.ceos22.cgv_clone.domain.booking.eneity.BookingSeat;
+import com.ceos22.cgv_clone.domain.booking.entity.Booking;
+import com.ceos22.cgv_clone.domain.booking.entity.BookingSeat;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
