+++ src/main/java/com/ceos22/cgv_clone/domain/booking/service/BookingService.java
@@ -4,8 +4,8 @@ import com.ceos22.cgv_clone.domain.booking.dto.request.BookingRequestDto;
 import com.ceos22.cgv_clone.domain.booking.dto.response.BookingCancelResponseDto;
 import com.ceos22.cgv_clone.domain.booking.dto.response.BookingDetailResponseDto;
 import com.ceos22.cgv_clone.domain.booking.dto.response.BookingResponseDto;
-import com.ceos22.cgv_clone.domain.booking.eneity.Booking;
-import com.ceos22.cgv_clone.domain.booking.eneity.BookingSeat;
+import com.ceos22.cgv_clone.domain.booking.entity.Booking;
+import com.ceos22.cgv_clone.domain.booking.entity.BookingSeat;
 import com.ceos22.cgv_clone.domain.booking.repository.BookingRepository;
 import com.ceos22.cgv_clone.domain.booking.repository.BookingSeatRepository;
 import com.ceos22.cgv_clone.domain.member.entity.Member;
